refactor(api): Extract 23 DTOs from conversations.rs to conversations_types.rs
Iteration 28: Apply sibling file pattern to conversations handler

- Create conversations_types.rs with 23 DTOs and helper functions (523 lines)
- Add 12 unit tests for DTO serialization/defaults
- Update mod.rs with conversations_types export
- Reduce conversations.rs from 1,197 to 812 lines (-32%)

Types extracted: ListConversationsParams, ListMessagesParams, ConversationResponse,
MessageResponse, FolderResponse, PaginatedConversationsResponse, PaginatedMessagesResponse,
PaginationMetaResponse, ConversationWithMessagesResponse, CreateConversationApiRequest,
UpdateConversationApiRequest, CreateMessageApiRequest, UpdateMessageApiRequest,
CreateFolderApiRequest, UpdateFolderApiRequest, BulkOperationRequest, BulkArchiveRequest,
BulkMoveRequest, BulkOperationResponse, ImportConversationsRequest, ImportConversationsResponse,
ImportErrorResponse, ShareResponse

All 222 tests pass (12 new DTO tests added).
